requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-taggit==3.0.0
django-user-visit==0.5.1
Pillow==9.2.0
sqlparse==0.4.2
ua-parser==0.15.0
user-agents==2.2.0
